+++ buildchain/buildchain/versions.py
@@ -19,7 +19,7 @@ Image = namedtuple("Image", ("name", "version", "digest"))
 
 CALICO_VERSION: str = "3.17.0"
 K8S_VERSION: str = "1.20.4"
-SALT_VERSION: str = "3002.2"
+SALT_VERSION: str = "3002.5"
 CONTAINERD_VERSION: str = "1.4.3"
 
 CALICO_RELEASE: str = "1"