+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-## Release 2.7.3 (in development)
+## Release 2.7.4 (in development)
+
+## Release 2.7.3
 ### Enhancements
 - [#2992](https://github.com/scality/metalk8s/issues/2992) - Check for conflicting
   packages (`docker`, `docker-ce` and `containerd.io`) on target machines before